move fragments to one directory
