+++ app/src/main/java/ru/kpfu/itis/homework2/Fragments/MusicFragment.kt
@@ -1,4 +1,4 @@
-package ru.kpfu.itis.homework2
+package ru.kpfu.itis.homework2.Fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,6 +6,9 @@ import android.view.View
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.bumptech.glide.Glide
+import ru.kpfu.itis.homework2.MusicAdapter
+import ru.kpfu.itis.homework2.MusicRepository
+import ru.kpfu.itis.homework2.R
 import ru.kpfu.itis.homework2.databinding.FragmentMusicBinding
