+++ app/src/main/java/ru/kpfu/itis/homework2/Fragments/ProfileFragment.kt
@@ -1,0 +1,7 @@
+package ru.kpfu.itis.homework2.Fragments
+
+import androidx.fragment.app.Fragment
+import ru.kpfu.itis.homework2.R
+
+
+class ProfileFragment : Fragment(R.layout.fragment_profile)
