+++ app/src/main/java/ru/kpfu/itis/homework2/Fragments/SettingsFragment.kt
@@ -1,0 +1,6 @@
+package ru.kpfu.itis.homework2.Fragments
+
+import androidx.fragment.app.Fragment
+import ru.kpfu.itis.homework2.R
+
+class SettingsFragment : Fragment(R.layout.fragment_settings)
