+++ app/src/main/java/ru/kpfu/itis/homework2/Fragments/VideoFragment.kt
@@ -1,0 +1,6 @@
+package ru.kpfu.itis.homework2.Fragments
+
+import androidx.fragment.app.Fragment
+import ru.kpfu.itis.homework2.R
+
+class VideoFragment : Fragment(R.layout.fragment_video)
