--- app/src/main/java/ru/kpfu/itis/homework2/ProfileFragment.kt
@@ -1,10 +1,0 @@
-package ru.kpfu.itis.homework2
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-
-class ProfileFragment : Fragment(R.layout.fragment_profile)
