--- app/src/main/java/ru/kpfu/itis/homework2/SettingsFragment.kt
@@ -1,9 +1,0 @@
-package ru.kpfu.itis.homework2
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class SettingsFragment : Fragment(R.layout.fragment_settings)
